Ensure order is valid on summary page
