+++ spec/features/order_spec.rb
@@ -1,0 +1,146 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.feature "Service ordering" do
+  include OmniauthHelper
+
+
+  context "as logged in user" do
+
+    let(:user) { create(:user) }
+    let(:service) { create(:service) }
+    let(:offer) { create(:offer, service: service) }
+
+    before { checkin_sign_in_as(user) }
+
+    scenario "I see project_item service button" do
+      visit service_path(service)
+
+      expect(page).to have_text("Order")
+    end
+
+    scenario "I see order open acces service button" do
+      @open_access_service = create(:open_access_service)
+      visit service_path(@open_access_service)
+
+      expect(page).to have_text("Add to my services")
+    end
+
+    scenario "I can order service" do
+      # force offer creation
+      offer
+      visit service_path(service)
+
+      click_on "Order"
+
+      # Step 1
+      expect(page).to have_current_path(service_offers_path(service))
+      expect(page).to have_text(service.title)
+      expect(page).to have_selector(:link_or_button,
+                                    "Next", exact: true)
+
+      select offer.iid
+      click_on "Next"
+
+      # Step 2
+      expect(page).to have_current_path(service_configuration_path(service))
+      expect(page).to have_selector(:link_or_button,
+                                    "Next", exact: true)
+
+      select "Services"
+      click_on "Next"
+
+      # Step 3
+      expect(page).to have_current_path(service_summary_path(service))
+      expect(page).to have_selector(:link_or_button,
+                                    "Order", exact: true)
+
+      expect do
+        click_on "Order"
+      end.to change { ProjectItem.count }.by(1)
+      project_item = ProjectItem.last
+      expect(project_item.offer_id).to eq(offer.id)
+
+      # Summary
+      expect(page).to have_current_path(service_summary_path(service))
+      expect(page).to have_selector(:link_or_button,
+                                    "Go to your service request", exact: true)
+
+      click_on "Go to your service request"
+
+      # Project item page
+      expect(page).to have_current_path(project_item_path(project_item))
+      expect(page).to have_content(service.title)
+    end
+
+    scenario "I'm redirected into service offers when offer is not chosen" do
+      visit service_configuration_path(service)
+
+      expect(page).to have_current_path(service_offers_path(service))
+    end
+
+    scenario "I'm redirected into service configuration when order is not valid" do
+      # Force offer creation
+      offer
+      visit service_offers_path(service)
+
+      # Select offer
+      select offer.iid
+      click_on "Next"
+
+      # Go directly to summary page
+      visit service_summary_path(service)
+
+      expect(page).to have_current_path(service_configuration_path(service))
+    end
+
+    scenario "I can order open acces service" do
+      open_access_service = create(:open_access_service)
+
+      visit service_path(open_access_service)
+
+      click_on "Add to my services"
+
+      expect(page).to have_current_path(service_offers_path(open_access_service))
+      expect(page).to have_text(open_access_service.title)
+      expect(page).to have_selector(:link_or_button,
+                                    "Next", exact: true)
+    end
+  end
+
+  context "as anonymous user" do
+    scenario "I nead to login to order service" do
+      service = create(:service)
+      user = create(:user)
+
+      visit service_path(service)
+
+      expect(page).to have_selector(:link_or_button, "Order", exact: true)
+
+      click_on "Order"
+
+      checkin_sign_in_as(user)
+
+      expect(page).to have_current_path(service_offers_path(service))
+      expect(page).to have_text(service.title)
+    end
+
+    scenario "I can see order button" do
+      service = create(:service)
+
+      visit service_path(service)
+
+      expect(page).to have_selector(:link_or_button, "Order", exact: true)
+    end
+
+    scenario "I can see openaccess service order button" do
+      open_access_service =  create(:open_access_service)
+
+      visit service_path(open_access_service)
+
+      expect(page).to have_selector(:link_or_button, "Add to my services", exact: true)
+      expect(page).to have_selector(:link_or_button, "Go to the service", exact: true)
+    end
+  end
+end
